1. many, many, many warnings fixed (from HP-UX build log; 50% are still left) 2. attempt (failed) at compiling wxGTK with GTK+ 1.3.0
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@5152 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/gifdecod.cpp
@@ -588,7 +588,7 @@ int wxGIFDecoder::ReadGIF()
             ppimg = &pimg->next;
 
             /* allocate memory for image and palette */
-            pimg->p   = (unsigned char *) malloc(size);
+            pimg->p   = (unsigned char *) malloc((size_t)size);
             pimg->pal = (unsigned char *) malloc(768);
 
             if ((!pimg->p) || (!pimg->pal))
